[Silver III] Title: 주유소, Time: 332 ms, Memory: 30236 KB -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -24,7 +24,7 @@ class Main{
 			for(int i =1; i<n-1; i++) {
 				price = Integer.parseInt(st.nextToken());
 				min = min > price ? price : min;
-				cost += min*distance[i+1];
+				cost += (long)min*distance[i+1];
 			}
 			
 			bw.write(cost+"");
